requirements.txt
asgiref==3.10.0
black==25.9.0
cfgv==3.4.0
click==8.3.0
coverage==7.10.7
distlib==0.4.0
Django==5.2.7
django-extensions==4.1
djangorestframework==3.16.1
drf-yasg==1.21.11
factory_boy==3.3.3
Faker==37.8.0
filelock==3.19.1
flake8==7.3.0
gunicorn==23.0.0
identify==2.6.15
inflection==0.5.1
iniconfig==2.1.0
isort==6.1.0
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dotenv==1.0.1
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
virtualenv==20.34.0
